Fixed menu to appear on all pages move messagelist to component folder for continuity todo: add condintional rendering in menu for changing menu-options
