+++ src/components/userCard/UserCard.js
@@ -9,25 +9,9 @@ class UserCard extends React.Component {
   constructor(props) {
     super(props);
     this.client = new DataService();
-    this.state = {
-      user: {
-        username: props.username,
-        displayName: props.displayName,
-        about: props.about,
-      },
-      statusCode: 0,
-    };
   }
 
-  componentDidMount() {
-    this.setState({
-      user: {
-        username: this.props.username,
-        displayName: this.props.displayName,
-        about: this.props.about,
-      },
-    });
-  }
+  componentDidMount() {}
   render() {
     if (!this.props.username) {
       return <div></div>;
@@ -35,11 +19,7 @@ class UserCard extends React.Component {
       return (
         <div className="UserCard">
           <GetUserPhoto username={this.props.username} />
-          <Card
-            title="All About Me!"
-            extra={<a href="/ProfileOptions">Edit</a>}
-            style={{ width: 300 }}
-          >
+          <Card title="User Information" style={{ width: 300 }}>
             <strong>Display Name: </strong>
             <h3>{this.props.displayName}</h3>
             <br />
