+++ src/pages/Home.js
@@ -13,8 +13,8 @@ class Home extends React.Component {
   render() {
     return (
       <div className="Home">
-        <Menu />
-        <h2>Your favorite microblogging platform</h2>
+        <Menu isAuthenticated={this.props.isAuthenticated} />
+        <h2>Home of Lady Lucky</h2>
         <LoginForm />
         <br />
         <button className="RegisterButton">
